% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifelseimage.R
\name{IfElseImage}
\alias{IfElseImage}
\title{IfElseImage}
\usage{
IfElseImage(condition,
  true.image = "https://dl.dropboxusercontent.com/u/539177224/uparrow_grey.svg",
  false.image = "https://dl.dropboxusercontent.com/u/539177224/downarrow_grey.svg")
}
\arguments{
\item{condition}{Expression evaluating to \code{TRUE} or {FALSE} that determines which image is displayed.}

\item{true.image}{URL to image (jpeg, png or svg) to be shown when \code{condition} is \code{TRUE}.}

\item{false.image}{URL to image (jpeg, png or svg) to be shown when \code{condition} is \code{FALSE}.}
}
\description{
Conditionally shows an image as a htmlwidget
}
\examples{
IfElseImage(TRUE)
IfElseImage(3 < 2,
            "https://dl.dropboxusercontent.com/u/539177224/thumbsup_grey.svg",
            "https://dl.dropboxusercontent.com/u/539177224/thumbsdown_grey.svg")
}
