% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outofkstars.R
\name{OutOfKStars}
\alias{OutOfKStars}
\title{OutOfKStars}
\usage{
OutOfKStars(x, number.rows, number.columns, auto.size = FALSE, width = 25)
}
\arguments{
\item{x}{Number of filled stars (out of \code{K = number.rows * number.columns})}

\item{number.rows}{The number of rows of stars.}

\item{number.columns}{The number of columns of stars.}

\item{auto.size}{Automatically sizes the plot based on the size of the window/slot.}

\item{width}{Width of a single star in pixels when \code{auto.size} is FALSE.}
}
\description{
OutOfKStars
}

