% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictochart.R
\name{PictoChart}
\alias{PictoChart}
\title{PictoChart}
\usage{
PictoChart(x, variable.image, base.image = "", K = max(ceiling(x)),
  direction = "horizontal", show.lines = FALSE, show.legend = FALSE,
  legend.text = "", icon.nrow = 1, icon.ncol = unlist(K)/icon.nrow,
  icon.fixedsize = FALSE, icon.halign = "left", icon.valign = "center",
  label.left = c(), label.right = c(), label.top = c(),
  label.bottom = c(), label.font = "arial", label.size = 10,
  label.weight = "normal", label.color = "#2C2C2C",
  label.top.font = label.font, label.top.size = label.size,
  label.top.weight = "400", label.top.color = label.color,
  label.top.height = 1.2 * label.top.size, label.right.font = label.font,
  label.right.size = label.size, label.right.weight = label.weight,
  label.right.color = label.color, label.right.width = 0,
  label.bottom.font = label.font, label.bottom.size = label.size,
  label.bottom.weight = label.weight, label.bottom.color = label.color,
  label.bottom.height = 1.2 * label.bottom.size,
  label.left.font = label.font, label.left.size = label.size,
  label.left.weight = label.weight, label.left.color = label.color,
  label.left.width = 0, label.left.halign = "right",
  label.right.halign = "left", label.top.halign = "center",
  label.bottom.halign = "left", label.left.valign = "center",
  label.right.valign = "center", label.top.valign = "center",
  label.bottom.valign = "center", legend.font = label.font,
  legend.size = 0.8 * label.size, legend.weight = label.weight,
  legend.color = label.color, row.height = 1.5 * label.size *
  max(icon.nrow), column.width = max(15 * max(icon.ncol), 0.5 * label.top.size
  * nchar(label.top), 0.5 * label.bottom.size * nchar(label.bottom)),
  bg.color = "transparent", line.color = "#A8A8A8", line.width = 0.5,
  pad.row = 5, pad.col = 5, pad.icon.row = 0, pad.icon.col = 0)
}
\arguments{
\item{x}{Data for charting}

\item{base.image}{Option URL of background image}

\item{K}{Maximum icons in each table cell. Ignored if both \code{icon.nrow} and \code{icon.ncol} supplied.}

\item{direction}{Accepts \code{horizontal}, \code{vertical}, \code{radial}, \code{scale}. (But \code{scale} may not be appropriate, especially if \code{K} is important).}

\item{icon.fixedsize}{When \code{true}, icons will not automatically resize to fill table cell.}

\item{icon.halign}{Horizontal alignment of icons in cell when \code{icon.fixedsize} is \code{true}. Accepts values of left, right or center both as scalar and vectors.}

\item{label.left}{Length must be equal to length (if \code{x} is a vector) or number of rows (if \code{x} is matrix or data.frame) as x. If no value is supplied, labels will be read from names/rowname of \code{x}. To suppress labels, use \code{label.left = rep("", length(x))}.}

\item{label.top}{By default, labels are read from column names of \code{x}.}

\item{label.font}{Controls font-family of all labels. To modify only the font of one label use \code{label.left.font, label.top.font}, etc.}

\item{row.height}{Height of graphic cells. Can be a single value or a numeric vector the same length as the number of rows in \code{x}.}

\item{column.width}{Width of graphic cells.}

\item{pad.row}{Single numeric specifying spacing between graphic cells in the table.}

\item{pad.icon.row}{Numeric specifying spacing between icons inside each table cell. May be a single value or a numeric matrix of the same dimensions as \code{x}.}

\item{pad.icon.ncol}{}
}
\description{
Function to create a chart of Pictographs
}

