% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictostdchart.R
\name{PictoStdChart}
\alias{PictoStdChart}
\title{Wrapper function to create Pictographs}
\usage{
PictoStdChart(x, variable.image, base.image = NULL, K = 0,
  read.KfromX = FALSE, groupBy = NULL, aggregate.period = "month",
  icon.autosize = FALSE, icon.halign = "left", direction = "horizontal",
  transpose = TRUE, show.labels = TRUE, show.table = FALSE)
}
\arguments{
\item{x}{Data to plot. Can be vector, matrix or data.frame.}

\item{variable.image}{URL of icon image.}

\item{base.image}{URL of base image (optional).}

\item{K}{Maximum number of icons in each table cell. This variable is ignored if \code{read.KfromX} is \code{true}.}

\item{groupBy}{Optional vector for aggregating \code{x}.}

\item{icon.autosize}{Rescale icons to fill up table cell}

\item{icon.halign}{Alignment of icons when \code{icon.autosize} is \code{false}. Should be one of left, center or right.}

\item{direction}{Direction in which icons are filled (\code{horizontal} or \code{vertical}).}

\item{transpose}{Transpose \code{x}.}

\item{show.labels}{Use default labels based on row/column names.}

\item{show.table}{Put table lines around row/column labels.}

\item{read.KfromX.}{If set to true, maximum number of icons in each row is taken from the last column of X. This option cannot be used with \code{groupBy}.}
}
\description{
\code{PictoStdChart} provides a simpler interface to create pictographs. In particular it can automatically aggregate data to different time scales.
}
\seealso{
\code{PictoChart()} provides finer control over plot parameters
}

