% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapesstdchart.R
\name{ShapesStdChart}
\alias{ShapesStdChart}
\title{Wrapper function to create Pictographs}
\usage{
ShapesStdChart(x, groupBy = NULL, variable.image = "red",
  gradient.type = "color", gradient.direction = "column",
  base.image = "grey80", image.type = "circle", direction = "radial",
  text.type = "percentage", text.position = "overlay", text.text = NULL,
  aggregate.period = "month", transpose = FALSE, hide.label.left = FALSE,
  hide.label.right = FALSE, hide.label.bottom = FALSE,
  hide.label.top = FALSE, label.left = c(), label.top = c(),
  label.bottom = c(), label.bottom.halign = "center", column.width = 20,
  row.height = 20, ...)
}
\arguments{
\item{x}{Data to plot. Can be vector, matrix or data.frame.}

\item{groupBy}{Optional vector for aggregating \code{x}.}

\item{direction}{Direction in which icons are filled (\code{horizontal}(default) or \code{vertical}). When vertical is used, the icons are placed in a single column per entry.}

\item{aggregate.period}{Time period in groupBy (e.g. "month", "year").}

\item{transpose}{Swap rows and columns in data matrix \code{x}.}

\item{hide.label.left}{Suppress labels on left of graphics. By default, if \code{label.left} is not supplied, it is taken from the rownames of \code{x}.}

\item{hide.label.top}{Suppress labels above graphics.}

\item{...}{Arguments passed to \code{PictoChart()}.}

\item{image}{Name of icon, e.g. \code{"star", "people",...}.}

\item{hide.base.image}{Turns off background image (on by default). In most cases it is appropriate to turn off the base image if K varies between entries.}
}
\description{
\code{ShapesStdChart}
}

