% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starbarchart.R
\name{StarBarChart}
\alias{StarBarChart}
\title{StarBarChart}
\usage{
StarBarChart(x, K, x.labels = NULL, label.font = "arial", label.size = 10,
  label.weight = "normal", label.color = "#2C2C2C")
}
\arguments{
\item{x}{Number of filled stars.}

\item{K}{Number of stars in each row.}
}
\description{
StarBarChart
}

