% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizevariable.R
\name{SummarizeVariable}
\alias{SummarizeVariable}
\title{Summarizes a numeric or categorical variable for plotting}
\usage{
SummarizeVariable(x, type = c("Average", "Sum", "Percentage")[1],
  weights = NULL, subset = NULL, category = NULL)
}
\arguments{
\item{x}{The input variable to summarize}

\item{type}{The type of summary statistic to output. This should be one of 'Average', 'Sum', 'Most frequent' or 'Percentage'.}

\item{weights}{Sampling or replication weights. This is a numeric vector of the same length as \code{x}.}

\item{subset}{Option logical vector of the same length as \code{x} indicating whether or not to include
an observation in \code{x}.}

\item{category}{A comma-seperated list of the name or indices of the categories to include for 'Percentage'.}
}
\description{
Summarizes a numeric or categorical variable for plotting
}
