% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictographchart.R, R/pictostdchart.R
\name{PictographChart}
\alias{PictoStdChart}
\alias{PictographChart}
\title{Create chart of pictographs}
\usage{
PictographChart(x, image = "star", image.type = "url",
  hide.base.image = FALSE, total.icons = NA, scale = NA, mode = "table",
  icon.palette = "Set1", icon.colors = "black",
  fill.direction = "fromleft", show.lines = FALSE, layout = NA,
  icon.nrow = 1, icon.ncol = NA, table.by.row = FALSE, bar.order = NA,
  column.order = NA, show.legend = FALSE, legend.text = "",
  legend.icon.color = NA, pad.legend = 40, hide.label.right = TRUE,
  hide.label.left = !hide.label.right, hide.label.bottom = (mode !=
  "column"), hide.label.top = (mode == "column"),
  label.color.asIcon = FALSE, label.left = NA, label.top = NA,
  label.bottom = NA, label.right = NA, sublabel.left = NA,
  sublabel.right = NA, label.pad = 0, label.left.pad = 0,
  label.right.pad = 0, label.bottom.align.horizontal = "center",
  label.left.align.horizontal = "center",
  label.right.align.horizontal = "center",
  label.top.align.horizontal = "center",
  label.left.align.vertical = "center", label.top.align.vertical = "center",
  label.right.align.vertical = "center",
  label.bottom.align.vertical = "center",
  sublabel.left.align.horizontal = "center",
  sublabel.right.align.horizontal = "center", width.height.ratio = NA,
  label.width = NA, label.top.height = NA, label.left.width = label.width,
  label.right.width = label.width, label.font.family = "arial",
  label.font.size = 12, label.font.color = "#2C2C2C",
  label.left.font.color = label.font.color,
  label.right.font.color = label.font.color,
  label.top.font.color = label.font.color,
  label.bottom.font.color = label.font.color,
  label.left.font.size = label.font.size,
  label.top.font.size = label.font.size,
  label.right.font.size = label.font.size,
  label.bottom.font.size = label.font.size,
  sublabel.left.font.size = label.left.font.size,
  sublabel.right.font.size = label.right.font.size,
  label.left.font.weight = "normal", label.top.font.weight = "normal",
  label.bottom.font.weight = "normal", label.right.font.weight = "normal",
  sublabel.left.font.weight = label.left.font.weight,
  sublabel.right.font.weight = label.right.font.weight,
  label.data.type = "none", label.data.text = NULL,
  label.data.position = ifelse(mode == "bar", "right", "footer"),
  label.data.font.weight = "normal", label.data.font.size = 0.8 *
  label.font.size, label.data.font.color = label.font.color,
  label.data.align.horizontal = "default", data.above.label = FALSE, ...)

PictoStdChart(...)
}
\arguments{
\item{x}{Data to plot. Can be vector, matrix or data.frame.}

\item{image}{Name of icon, e.g. \code{"star", "stickman",...}.}

\item{hide.base.image}{Turns off background image (on by default). In general, the base image should only be shown if the input data is a proportion.}

\item{total.icons}{Maximum number of icons in each table cell. By default, it will be taken to be \code{ceiling(x)} (if icon.autosize is on) or \code{ceiling(max(x))}. This variable is ignored if \code{read.KfromX} is \code{true}.}

\item{scale}{Value of one icon. If \code{scale  =  0}, the value is automatically determined from the data so that the largest entry is represented by 10 icons.}

\item{mode}{Can be set to one of \code{"table", "bar", "column"}. For options \code{bar} and \code{column}, the chart is constrained to look like a bar or column chart. e.g For \code{mode  =  "column"}, 1-dimensional vectors/matrices are re-shaped to have multiple columns, labels are put below the graphis and icons are arranged vertically. Option \code{mode  =  "table"} is the most general and does not impose constraints.}

\item{fill.direction}{Direction in which icons are filled (\code{horizontal}(default) or \code{vertical}). When vertical is used, the icons are placed in a single column per entry.}

\item{bar.order}{Specify order of bars when \code{mode=="bar"}. Should be one of \code{NA, "Ascending", "Descending"}.}

\item{column.order}{Specify order of bars when \code{mode=="column"}. Should be one of \code{NA, "Ascending", "Descending"}.}

\item{legend.text}{Text shown with legend. If this string is empty, it will be automatically filled in using \code{scale}. (To hide text completely, use \code{legend.text  =  " "})}

\item{hide.label.left}{Suppress labels on left of graphics. By default, if \code{label.left} is not supplied, it is taken from the rownames of \code{x}.}

\item{hide.label.top}{Suppress labels above graphics.}

\item{...}{Arguments passed to \code{PictoChart()}.}

\item{data.label.position}{When \code{label.data.type != "none"}, the position of the data labels can be one of \code{"Above icons", "Below icons"} (all modes) or \code{"On left", "On right"} (bar mode only). Note that \code{"On left"} will overrride \code{sublabel.left} and \code{"On right"} will overrride \code{sublabel.right}.}
}
\description{
\code{PictographChart} Create a chart to visually represent the data in a table or vector using the number of icons displayed

\code{PictoStdChart} Deprecated function for backwards-compatibility
}
\seealso{
PictoChart
}

