% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlepicto.R
\name{SinglePicto}
\alias{SinglePicto}
\title{SinglePicto}
\usage{
SinglePicto(x, number.rows, number.columns, image = "star",
  hide.base.image = FALSE, direction = "horizontal", auto.size = FALSE,
  width = 50, bg.color = "transparent", text.type = "none",
  text.size = 8, text.weight = 700, text.halign = "center")
}
\arguments{
\item{x}{Number of filled icons (out of \code{K = number.rows * number.columns})}

\item{number.rows}{The number of rows of icons.}

\item{number.columns}{The number of columns of stars.}

\item{image}{name of icon}

\item{hide.base.image}{Set to \code{TRUE} to use blank background instead of background image.}

\item{direction}{Direction in which pictograph is filled (horizontal, vertical or radial).}

\item{auto.size}{Automatically sizes the plot based on the size of the window/slot.}

\item{width}{Width of a single icon in pixels when \code{auto.size} is FALSE.}

\item{bg.color}{Background color of graphic.}

\item{text.type}{One of \code{"none", "count", "proportion", "percentage"}}
}
\description{
Creates a single pictograph. Allows customization of the number of icons
and dimensions.
}
\seealso{
PictoStdChart to create a chart or table of pictographs
}

