% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlepicto.R
\name{SinglePicto}
\alias{SinglePicto}
\title{SinglePicto}
\usage{
SinglePicto(x, number.rows, number.columns, image = "star.filled",
  base.image = "star.empty", direction = "horizontal", auto.size = FALSE,
  width = 25)
}
\arguments{
\item{x}{Number of filled icons (out of \code{K = number.rows * number.columns})}

\item{number.rows}{The number of rows of icons.}

\item{number.columns}{The number of columns of stars.}

\item{image}{name of icon}

\item{base.image}{name of background icon. Set to "none" for no background.}

\item{direction}{Direction in which pictograph is filled (horizontal, vertical or radial).}

\item{auto.size}{Automatically sizes the plot based on the size of the window/slot.}

\item{width}{Width of a single star in pixels when \code{auto.size} is FALSE.}
}
\description{
Creates a single pictograph. Allows customization of the number of icons
and dimensions.
}
\seealso{
PictoStdChart to create a chart or table of pictographs
}

