% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlepicto.R
\name{SinglePicto}
\alias{SinglePicto}
\title{SinglePicto}
\usage{
SinglePicto(x, total.icons = NA, image = "star", base.image = "",
  is.custom.url = FALSE, number.rows = NA, number.cols = NA,
  width.height.ratio = 1, graphic.width = NA, graphic.height = NA,
  layout = NA, scale = 1, hide.base.image = FALSE,
  fill.direction = "fromleft", fill.icon.color = "black",
  base.icon.color = "", background.color = "transparent",
  auto.size = FALSE, icon.width = 50, pad.row = 0, pad.col = 0,
  margin = 0, margin.top = margin, margin.right = margin,
  margin.bottom = margin, margin.left = margin,
  label.data.position = c("None", "Above", "Below")[1],
  label.data.digits = 0, label.data.bigmark = ",", label.data.prefix = "",
  label.data.suffix = "", label.data.100prc = FALSE,
  label.data.font.weight = "normal", label.data.font.size = 12,
  label.data.font.family = "arial", label.data.font.color = "#2C2C2C",
  label.data.align.horizontal = "center", graphic.resolution = 96,
  font.whratio = 0.6, print.config = FALSE, x.limit = 1000)
}
\arguments{
\item{x}{Input data which determines the number of icons (\code{x/scale}) filled in the pictograph.}

\item{total.icons}{Total number of icons. Defaults to \code{total.icons=ceiling(x/scale)}.}

\item{image}{The name of the icon to use (e.g. \code{"star", "stickman"}) or the URL of an image when \code{is.custom.url} is true.}

\item{base.image}{The URL of the base image. Only used if \code{is.custom.url = TRUE} and \code{hide.base.image = FALSE}.}

\item{is.custom.url}{When set to true, image is expected to be a URL to an jpeg or png image file available from online.}

\item{number.rows}{If neither \code{number.rows} and \code{number.cols} is supplied, the default behaviour is to place icons according to \code{width.height.ratio}. Note that number.rows is ignored when number.cols is non-zero.}

\item{number.cols}{Maximum number of icons in each column. Overrides \code{number.rows} and \code{width.height.ratio}.}

\item{width.height.ratio}{Width to height ratio of pictograph if \code{layout == "Width-to-height ratio"}.}

\item{layout}{Optional parameter to determine how the layout is specified. Can be one of \code{"Width-to-height ratio", "Number of rows", "Number of columns", "Fill graphic"}. If not supplied, a decision will be made based on which parameters are supplied}

\item{scale}{Scaling factor for \code{x}. Defaults to 1.}

\item{hide.base.image}{Set to \code{TRUE} to use blank background instead of base image.}

\item{fill.direction}{Direction in which pictograph is filled (one of \code{"fromleft","fromright","fromtop","frombottom"}).}

\item{fill.icon.color}{Color of the filled icons. Only applicable for built-in icons.}

\item{base.icon.color}{Color of the unfilled icons when \code{hide.base.image == FALSE}. Defaults to grey (#CCCCCC). Only applicable for built-in icons.}

\item{background.color}{Color of the graphic background}

\item{auto.size}{Automatically sizes the plot based on the size of the window/slot.}

\item{icon.width}{Width of a single icon in pixels when \code{auto.size} is \code{FALSE}.}

\item{pad.row}{Vertical space between icons. This should be a number between 0 (no space) and 1.0 (all space).}

\item{pad.col}{Horizontal space between icons.}

\item{margin}{Controls space on margins of the graphic. When \code{margin} is used, space on all 4 sides are adjusted simultaneously, but margins can also be adjusted separately using \code{margin.top, margin.right, margin.bottom, margin.left}.}

\item{label.data.position}{One of \code{"None"}, \code{"Above"} or \code{"Below"}.}

\item{label.data.digits}{Number of digits of the data label to show.}

\item{label.data.bigmark}{Option to prettify large numbers. By default a comma is placed after a thousand.}

\item{label.data.prefix}{String to prepend data label.}

\item{label.data.suffix}{String to append to data label.}

\item{label.data.100prc}{Option to show data labels multiplied by 100. This is useful when reporting percentages.}

\item{print.config}{If set to \code{TRUE}, the JSON string used to generate pictograph will be printed to standard output. This is useful for debugging.}

\item{x.limit}{Upper limit of x above which \code{scale} is automatically calculated. This can be set to \code{NA}, but may cause slowness or freezing when the user inputs a large \code{x}.}
}
\description{
Creates a single pictograph. Allows customization of the number of icons
and dimensions.
}
\examples{
xx <- 4
SinglePicto(xx)
SinglePicto(xx, total.icons=10, image="stickman", number.cols=5,
   fill.icon.color="red", base.icon.color="deepskyblue")
SinglePicto(xx, 9, number.rows=3, is.custom.url=TRUE,
   image="http://wiki.q-researchsoftware.com/images/9/91/Star_filled.svg",
   base.image="http://wiki.q-researchsoftware.com/images/2/21/Star_unfilled.png")
}
\seealso{
PictographChart to create a chart or table of pictographs
}

